requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
dataclasses-json==0.6.7
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1758162063/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
Django==4.2.16
django-cors-headers==4.9.0
djangorestframework==3.16.1
etelemetry==0.3.1
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
faiss==1.8.0
faiss-cpu==1.8.0
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1761567966/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1762350976/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.63
langchain-google-genai==2.0.4
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.8
langsmith==0.1.147
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.7
llama-index-cli==0.5.3
llama-index-core==0.14.7
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-gemini==0.6.1
llama-index-llms-openai==0.6.7
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-faiss==0.5.1
llama-index-workflows==2.11.1
llama-parse==0.6.54
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.2
numpy @ file:///D:/bld/numpy_1707225528937/work/dist/numpy-1.26.4-cp311-cp311-win_amd64.whl#sha256=537f0573fefb1e93b3633e8a79102e6b6d2911e09e0a90008993efef4242448e
openai==1.109.1
orjson==3.11.4
packaging==24.2
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1759953252/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psutil @ file:///D:/bld/bld/rattler-build_psutil_1762092931/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydot==4.0.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PyMuPDF==1.26.6
pymupdf4llm==0.1.9
pyparsing==3.2.5
pypdf==6.1.3
pypdfium2==5.0.0
pytesseract==0.3.13
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487145/work
pyxnat==1.6.4
PyYAML==6.0.3
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387035/work
rdflib==7.4.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==3.1.1
simplejson==3.20.2
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
tornado @ file:///D:/bld/tornado_1762506678249/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
traits==7.0.2
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
wrapt==1.17.3
yarl==1.22.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
